@pschafhalter 2018-06-22
Moved to modin-project/modin#4.